Add base application structure
Solution and the two main projects (Sculptor and Sculptor.Core) have
been created. This structure is subject to change, but for now gives a
rough idea on how concerns should be separated.

There is currently only one 'Command' and it does nothing other than
print out a success message to the console - but it does demonstrate
program flow quite nicely.