Refactor the way commands are validated
Validation has now been moved out of the CommandHandlers and into a
dedicated component that is solely responsible for ensuring invalid
data does not reach the inner logic. This allows us to make assumptions
about the state of the objects we're working with and reduces
unecessary rechecking as the Handler logic will only be invoked if
all validation rules have passed.

This is achieved via a new Nuget package (`FluentValidation`) and a
decorator (`ValidationCommandHandlerDecorator.cs`) which wraps every
CommandHandler via the registration line in `Bootstrapper.cs`.

To assist with this change, the way in which errors are reported to the
user has been standardised against the format used by the `CommandLine`
Library. This keeps user experience consistent between errors raised
parsing the commands, and validation errors raised evaluating the
commands.

Other minor changes made in relation to the above include:
* New project (`Sculptor.Infrastructure`) to contain dependencies
common to both `Sculptor` and `Sculptor.Core`.
* The `IUserInput` interface now provides access to the original
command submitted by the user so that it's available for future
processing.
* Removal of all unit tests until the design has stabilised for
managing command parsing, validation etc as nothing is final yet.