Refactor towards a feature folder convention
Moved the core files around to follow a feature folder convention,
whereby every new command has all of its related files scoped to a
single directory. Using the `CreateCommand` as an example we now have
the following structure:

| Sculptor.Core
    | Domain
        | Create
            ~ CreateCommand.cs
            ~ CreateCommandHandler.cs
            ~ CreateCommandValidator.cs
        ~ ICommand.cs
        ~ ICommandHandler.cs

This should make it easier to find related logic as the project grows
in size with more commands. There may be further tweaks to the exact
hierarchy shown above, but this is the kind of structure I'd like to
implement for all areas of the program for the sake of maintainability.