Refactor `Application.cs` so that it's testable
Access modifiers changed so that the `Application` class is
discoverable by the `Sculptor.Tests` project. This has allowed us to
verify that the `CommandParser` dependency is correctly invoked when
the `Application.Run()` method is called.

* The `SplitArguments()` method previously found in the `Program.cs`
file has been relocated to the `Sculptor.Infrastructure` project under
a new static helper class: `ArgumentHelper`.
This allows easier access when debugging or writing tests that target
the initial startup classes as you can just provide any command in the
same manner you would running the application in the terminal.
* Miscellaneous `TODO` comments added to outline the next batch of
changes required.