Incorporate configuration controlled logging
In the interests of providing a configuration file to the end user so
that they can control some aspects of the Sculptor tool we now have an
`appsettings.json` file that is copied to the output directory on
build.

Currently this file only controls the log settings exposed by the new
dependency we have on the `Serilog` Nuget package. The process is
pretty consistent with the library documentation except for how we set
the path format i.e:

`"pathFormat": "<<PATH_FORMAT>>"`

Because we need to have a default, and we won't know the exact value
until runtime (as this could be running on a number of different OSs)
we need to do a string replace that looks for the placeholder value
shown above. This is done via the `Bootstrapper.BuildConfiguration`
method and it works at the cost of feeling a bit "hacky". This will
need revisiting at a later date.

With this value updated as soon as the Bootstrapper has finished its
registrations we will have ensured that a directory exists on disk
for us to start writing messages to and a config file to control the
specifics that is re-parsed on each run.

Other adjustments made to facilitate this change include:
* Removal of all `Trace.Write()` method calls and replacement with the
appropriate level of log output.
* Verb Registration updated to return an `IEnumerable<Type>` rather
than a `Type[]` so that we can defer its evaluation until required.
* Unit tests updated to reflect new logging dependency and altered
"expected" output.
* Various logging statements added to try show examples of how
program execution should be recorded. This includes writing out as
much info as possible when we encounter an exception - both 'expected'
due to invalid data, and 'unexpected' which we can't recover from.