Run formatter to correct line endings and syntax
The entire solution has been run through the Code Maid extension for
Visual Studio so that the last set of commits (`bf332dd` to `24ca802`)
are consistent with the style rules for this project.

This was due to the same level of formatting tools not being available
on OSX (where the code for these commits were written) as part of
Visual Studio for Mac.

TODO: find a way of enforcing style rules across platforms.