Fix invalid paths being generated on Windows
The file path we generate and save to the `config.json` file for
controlling where log files are stored was previously being built via
the following method call:

`System.IO.Path.Combine(...,...)`

This worked as intended on nix based environments, but on Windows the
format of: `somefolder\subfolder\subsubfolder` was resulting in an
invalid JSON file being created (as these directory separators need to
be escaped).

This fix simply enforces the `System.IO.Path.AltDirectorySeparatorChar`
value to be used to separate directories instead of the default
`System.IO.Path.DirectorySeparatorChar`. For nix systems this will
not change the behaviour as both fields point to: '/'.

On Windows (and any other OS that uses a backslash as their default
separator) this will result in '\' being replaced with '/' which will
ensure a valid config value is always written.