Update startup logic to be more idiomatic
The previous `if DEBUG` statement overriding the `args[]` parameter
value in the `Main` method has been replaced in favour of using the
`Arguments` value exposed in Visual Studio for the Sculptor project
file. As this value is not persisted it will not become a source of
conflict across merges.

* Documentation also updated (via README.md file) to reflect currently
available commands / arguments.