Refactor parsing of `IUserInput` to `ICommand`
Reworked the application flow from receiving user input, through to
parsing it to an implementation of `ICommand` somewhere in our Domain
layer. This has resulted in better separation of concerns - for example,
we no longer expose the `ParserResult<object>` to all dependencies via
the `IUserInput` interface. Instead we only expose the text helper for
message formatting on the new `IOutputFormatter` interface.

In addition, exception handling has been reworked so that the
`Application` class is responsible for handling custom, "expected"
exceptions, and the `Program` class now only has to bother with the
truly unexpected fatal errors.

Custom exceptions can now implement the `IFormattableException`
interface to have logging and console output automatically wired up -
reducing the amount of duplicated code being written in exception
handlers. This interface also supports more complex exceptions that
require additional transformation before the user can be informed (see
the `Sculptor.Infrastructure.Exceptions.ValidationExceptions` namespace
for further detail).

* XML documentation added and improved to better explain the role of
each dependency.
* `IRegisteredVerbs` interface reverted to returning an array of types
as this value will not change during program execution so it's better
not to re-evaluate it on every read.