Refactor auditing and exception management
Through the introduction of the `AuditCommandHandlerDecorator` class all
Command Handlers will now be automatically wrapped. By doing this we
have a single point for controlling the recording of information like:

* What command(s) have been run
* How long a command took to complete
* Any other meta information we want to add in the future

In addition, exception handling has been refactored so that it's split
between the exceptions we "expect" from our domain (handled by the new
`ExceptionCommandHandlerDecorator` class), and those truly unexpected
that we can't recover from (or give the user much meaningful feedback).

This has the added bonus of reducing the `Program` class back to a
simple entry point - not needing to worry about concerns such as logging
exceptions raised etc.