Add basic hosting ability via Kestrel
There is now a new command available (`sculptor serve`) which when run
in a previously created Sculptor project's root directory will start a
local Kestrel webserver capable of serving static assets from the
configured output path.

In addition, as the serve command is currently simple (with only a
single optional argument) there is a new interface to represent concerns
such as this which do not expect a decorated validator instance (see the
usage of the `IExcludeFromValidation` interface).

Other minor changes as part of this feature include:

* A `launchSettings.json` file for holding command line arguments.
* Refactoring of `CommandScope` so that it's relevant only when running
in DEBUG configurations.
* Updating all references to `Environment.CurrentDirectory` to use the
new static `FilePathHelper.ExecutingDirectory` instead. This value can
then be freely changed (depending on the solution configuration) to
force Sculptor to run in a directory it wouldn't normally.
* Fixed a bug in the `ConfigComposer` class to ensure platform
independent paths are always used.