Clarify installation process and fix path bug
Updated the documentation to reflect the actual package and installation
steps necessary when building from source.

While documenting this a bug in the paths being used for running the
Sculptor tool was noticed. This was remedied by updating the
`FilePathHelper` to reflect the difference between the current executing
directory and the current working directory. Both can now be overriden
if and when necessary.